project1/requirements.txt
python-dotenv
requests
python-pptx
pydantic
pandas
Pillow
